+++ python3/perceptron/pyproject.toml
@@ -9,7 +9,7 @@ python = "^3.10"
 matplotlib = "^3.8.2"
 
 [tool.poetry.dev-dependencies]
-pytest = "^5.2"
+pytest = "^7.4.4"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
