+++ python3/perceptron/test
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+poetry run pytest
